Deprecate Spree::Deprecation in favor of Spree.deprecator
Use autoload to print a deprecation only if the constant is loaded.

(cherry picked from commit 848bf9f1381073c9d7c8351689158395c5851f05)

# Conflicts:
#	core/lib/spree/deprecation.rb
