+++ .rubocop.yml
@@ -341,7 +341,7 @@ Lint/SuppressedException:
 
 Lint/MissingSuper:
   Exclude:
-    - 'core/lib/spree/deprecation.rb' # this is a known class that doesn't require super
+    - 'core/lib/spree/deprecated_instance_variable_proxy.rb' # this is a known class that doesn't require super
     - 'core/lib/spree/preferences/configuration.rb' # this class has no superclass defining `self.inherited`
 
 Rails/FindEach:
