+++ core/lib/spree/preferences/preferable_class_methods.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 
-require 'spree/deprecation'
 require 'spree/encryptor'
 
 module Spree::Preferences
